refactor: chat 도메인 관련 api 함수를 entities 레이어로 이동 (#69)
