+++ src/entities/chat-room/api/api.ts
@@ -1,10 +1,5 @@
 import { api } from '@/shared/api/client'
 import {
-  ChatMessageListResponseSchema,
-  type ChatMessageListRequest,
-  type ChatMessageListResponse,
-} from '@/entities/message/model/schema'
-import {
   ChatRoomListResponseSchema,
   type ChatRoomListResponse,
   type ChatRoomEnterResponse,
@@ -22,16 +17,4 @@ export const enterChatRoom = async (
 ): Promise<ChatRoomEnterResponse> => {
   const response = await api.post(`/chat/${roomId}`)
   return response.data
-}
-
-// ---------- 채팅 메세지 조회 ----------
-export const getChatMessageList = async ({
-  roomId,
-  size = 20,
-  cursor,
-}: ChatMessageListRequest): Promise<ChatMessageListResponse> => {
-  const response = await api.get(`/chat/${roomId}/messages`, {
-    params: { cursor, size },
-  })
-  return ChatMessageListResponseSchema.parse(response.data)
 }
